media destroy route

+++ app/Http/Controllers/Api/MediaAlbumController.php
@@ -7,13 +7,17 @@ use App\Http\Requests\Api\IndexMediaAlbumRequest;
 use App\Http\Requests\Api\StoreMediaAlbumRequest;
 use App\Http\Resources\MediaAlbumResource;
 use App\Http\Resources\MediaAlbumsResource;
+use App\Models\Media;
 use App\Models\MediaAlbum;
+use Illuminate\Support\Facades\Gate;
 
 class MediaAlbumController extends Controller
 {
     public function index(IndexMediaAlbumRequest $request)
     {
         $mediaAlbums = MediaAlbum::whereIn('id', $request->validated('album_ids'))->get();
+
+        $mediaAlbums->each(fn($album) => Gate::authorize('ownsAlbum', $album));
 
         return MediaAlbumsResource::collection($mediaAlbums);
     }
@@ -24,10 +28,11 @@ class MediaAlbumController extends Controller
             ['id' => $request->validated('id')],
             ['user_id' => auth()->id()]
         );
-        // TODO:: napraviti validaciju ownershipa
+
+        Gate::authorize('ownsAlbum', $mediaAlbum);
 
         $media = $request->safe()->collect('files')->map(fn($file) =>
-            $mediaAlbum->addMedia($file)->toMediaCollection()
+        $mediaAlbum->addMedia($file)->toMediaCollection()
         );
 
         return MediaAlbumResource::collection($media);
@@ -35,6 +40,8 @@ class MediaAlbumController extends Controller
 
     public function show(MediaAlbum $mediaAlbum)
     {
+        Gate::authorize('ownsAlbum', $mediaAlbum);
+
         return MediaAlbumResource::collection(
             $mediaAlbum->media()->paginate()
         );
@@ -42,8 +49,19 @@ class MediaAlbumController extends Controller
 
     public function destroy(MediaAlbum $mediaAlbum)
     {
+        Gate::authorize('ownsAlbum', $mediaAlbum);
+
         $mediaAlbum->media()->delete();
         $mediaAlbum->delete();
+
+        return response()->noContent(200);
+    }
+
+    public function destroyMedia(MediaAlbum $mediaAlbum, Media $media)
+    {
+        Gate::authorize('ownsAlbum', $mediaAlbum);
+
+        $media->delete();
 
         return response()->noContent(200);
     }
